Upgraded SingleBoard_MiniMax Bot.
I improved move selection time on an empty board by just randomly selection a tile. This is fine because the scores for an empty board will be all 0s so minimax will select a random tile anyways. I upgraded the bot so that when there is more then once active board it will now randomly select a active board instead o just selecting the first one it sees. Also I upgraded to move selection when there are ties for the scores so that it will randomly select between its best available moves.