Upgraded SingleBoard_MiniMax Bot
I added the ability to evaluate all the active boards instead of just a random one. The minimax algorithm will be run on every active board each turn and choose the board with the max returned value from the minimax algorithm. The bot always chooses to attack and win a game over blocking his opponent.